server/pyproject.toml
[project]
name = "kiwi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-allauth>=65.4.0",
    "django-cors-headers>=4.6.0",
    "django-crum>=0.7.9",
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "psycopg[binary]>=3.2.4",
    "uuid-utils>=0.10.0",
]
